Changes
Lots of changes are done to the code base, I should keep up with my
commits.

So whats done is I've made the commands the UX commands and not exactly
the FTP commands, the FTP commands are done behind the scenes.

I've changed mode to an enum for preperation for RFC 2428.

The code works perfectly and without errors atm.